Accountant View and email records
Created viewrecords.php, emailrecords.php, and recordinsert.php.  Which provide functionality for an accountant to view, sort and filter all employee records, as well as email the results to their inbox.

Also  updated many  top comments to be in php sections, so as not to display at the top of the html source.